requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
gTTS==2.3.2
python-dotenv==1.0.0
requests==2.32.3


# --- Core App ---
flask
flask-cors
python-dotenv

# --- LLM + Embeddings ---
openai>=1.10.0
sentence-transformers
torch
transformers

# --- Vector DB ---
chromadb>=0.5.0

# --- TTS ---
gTTS
pydub  # NOT used directly, but required by gTTS sometimes
ffmpeg-python  # wrapper, not mandatory but helps safe usage

# --- Utilities ---
numpy
pandas

# --- For Debugging & Dev ---
ipykernel
ftfy

